requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
BAScraper==0.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-default-group==1.2.4
duckdb==1.1.3
faiss-gpu==1.10.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
propcache==0.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
setuptools==75.8.0
six==1.17.0
sqlite-fts4==1.0.3
sqlite-migrate==0.1a2
sqlite-utils==3.38
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
urllib3==2.2.3
warcdb==0.2.2
warcio==1.7.4
yarl==1.18.3
